chore: Renamed mobile/ducks to mobile/modules
