+++ src/drive/mobile/modules/authorization/duck/actions.js
@@ -4,8 +4,8 @@ import {
   resetClient
 } from 'drive/mobile/lib/cozy-helper'
 import { resetPersistedState } from 'drive/store/persistedState'
-import { setUrl, setOffline } from 'drive/mobile/ducks/settings/duck'
-import { startReplication } from 'drive/mobile/ducks/replication/duck'
+import { setUrl, setOffline } from 'drive/mobile/modules/settings/duck'
+import { startReplication } from 'drive/mobile/modules/replication/duck'
 
 export const SET_TOKEN = 'SET_TOKEN'
 export const SET_CLIENT = 'SET_CLIENT'
