+++ src/drive/mobile/modules/mediaBackup/components/UploadProgression.jsx
