+++ src/drive/mobile/modules/mediaBackup/duck/index.js
@@ -9,7 +9,7 @@ import {
 } from 'drive/mobile/lib/media'
 import { isWifi } from 'drive/mobile/lib/network'
 import { logException } from 'drive/mobile/lib/reporter'
-import { setBackupImages } from 'drive/mobile/ducks/settings/duck'
+import { setBackupImages } from 'drive/mobile/modules/settings/duck'
 import {
   MEDIA_UPLOAD_START,
   MEDIA_UPLOAD_END,
