+++ src/drive/mobile/modules/onboarding/components/BackupPhotosVideos.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { connect } from 'react-redux'
-import { backupImages } from 'drive/mobile/ducks/mediaBackup/duck'
+import { backupImages } from 'drive/mobile/modules/mediaBackup/duck'
 import OnBoardingStep from './OnBoardingStep'
 
 export const BackupPhotosVideos = ({ onActivate, onSkip, breadcrumbs }) => (
