+++ src/drive/mobile/modules/onboarding/components/Wizard.jsx
