+++ src/drive/mobile/modules/replication/duck.js
@@ -2,8 +2,8 @@
 import { openFolder, getOpenedFolderId } from 'drive/actions'
 import { startReplication as startPouchReplication } from 'drive/mobile/lib/replication'
 import { resetClient } from 'drive/mobile/lib/cozy-helper'
-import { revokeClient as reduxRevokeClient } from 'drive/mobile/ducks/authorization/duck'
-import { isOfflineCapable } from 'drive/mobile/ducks/settings/duck'
+import { revokeClient as reduxRevokeClient } from 'drive/mobile/modules/authorization/duck'
+import { isOfflineCapable } from 'drive/mobile/modules/settings/duck'
 
 const SET_FIRST_REPLICATION = 'SET_FIRST_REPLICATION'
 const SET_POUCH_INDEXES = 'SET_POUCH_INDEXES'
