+++ src/drive/mobile/modules/settings/RatingModal.jsx
