+++ src/drive/mobile/modules/settings/duck/reducers.js
